get ready for large numbers

+++ e002/src/main.rs
@@ -1,9 +1,9 @@
 use std::{thread::sleep, time::Duration};
 
 fn main() {
-    const UPPER_BOUND: u64 = 4_000_000;
+    const UPPER_BOUND: u128 = 4_000_000;
 
-    let res: u64 = e_math::generate::fibonacci()
+    let res: u128 = e_math::generate::fibonacci()
         .take_while(|x| *x <= UPPER_BOUND)
         .filter(|x| x % 2 == 0)
         .sum();
